+++ examples/particles.rs
@@ -12,8 +12,8 @@ use coffee::input::{KeyCode, KeyboardAndMouse};
 use coffee::load::{loading_screen::ProgressBar, Join, Task};
 use coffee::{Game, Result, Timer};
 
-// Try increasing this value! I (@hecrj) can render 350k particles at 100fps
-// on my system. I have not tried going above that, yet...
+// Try increasing this value! I (@hecrj) can render 1 MILLION particles at
+// 90 fps on my system (4790k, GTX 980, Windows 7) using Vulkan.
 const AMOUNT: u32 = 50_000;
 
 // Play with these values to alter the way gravity works.
@@ -146,20 +146,19 @@ impl Game for Particles {
     fn draw(&mut self, state: &State, frame: &mut Frame, timer: &Timer) {
         frame.clear(Color::BLACK);
 
-        // Draw particles all at once!
-        self.batch.clear();
-
+        // When interpolating, we need to know how close the next tick is
         let delta_factor = if self.interpolate {
             timer.next_tick_proximity()
         } else {
             0.0
         };
 
-        for particle in &state.particles {
+        // Generate sprites in parallel! <3 rayon
+        let sprites = state.particles.par_iter().map(|particle| {
             let velocity =
                 particle.velocity + particle.acceleration * delta_factor;
 
-            self.batch.add(Sprite {
+            Sprite {
                 source: Rectangle {
                     x: Self::particle_color(velocity),
                     y: 0,
@@ -168,9 +167,16 @@ impl Game for Particles {
                 },
                 position: particle.position + velocity * delta_factor,
                 scale: (1.0, 1.0),
-            });
-        }
+            }
+        });
 
+        // Clear batch contents from previous frame
+        self.batch.clear();
+
+        // Use the parallel iterator to populate the batch efficiently
+        self.batch.par_extend(sprites);
+
+        // Draw particles all at once!
         self.batch
             .draw(Point::new(0.0, 0.0), &mut frame.as_target());